Z-Image-worker/requirements.txt
# Z-Image Worker dependencies
# Install with: pip install -r requirements.txt
# For PyTorch with CUDA 12.1 (recommended):
#   pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 --index-url https://download.pytorch.org/whl/cu121
# For PyTorch with CUDA 11.8:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

accelerate>=1.0.0
diffusers @ git+https://github.com/huggingface/diffusers@67613369bb92f120eff221fc571dc2901fe994f6
httpx>=0.27.0
huggingface_hub>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
safetensors>=0.4.0
sentencepiece>=0.1.99
tokenizers>=0.19.0
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.40.0
